+++ griptape/tasks/base_task.py
@@ -4,18 +4,19 @@ import logging
 import uuid
 from abc import ABC, abstractmethod
 from enum import Enum
+from functools import cached_property
 from typing import TYPE_CHECKING, Any, Optional
 
 from attrs import Factory, define, field
 
-from griptape.artifacts import ErrorArtifact
+from griptape.artifacts import BaseArtifact, ErrorArtifact
 from griptape.configs import Defaults
 from griptape.events import EventBus, FinishTaskEvent, StartTaskEvent
 from griptape.mixins.futures_executor_mixin import FuturesExecutorMixin
+from griptape.mixins.runnable_mixin import RunnableMixin
 from griptape.mixins.serializable_mixin import SerializableMixin
 
 if TYPE_CHECKING:
-    from griptape.artifacts import BaseArtifact
     from griptape.memory.meta import BaseMetaEntry
     from griptape.structures import Structure
 
@@ -23,7 +24,7 @@ logger = logging.getLogger(Defaults.logging_config.logger_name)
 
 
 @define
-class BaseTask(FuturesExecutorMixin, SerializableMixin, ABC):
+class BaseTask(FuturesExecutorMixin, SerializableMixin, RunnableMixin, ABC):
     class State(Enum):
         PENDING = 1
         EXECUTING = 2
@@ -53,7 +54,7 @@ class BaseTask(FuturesExecutorMixin, SerializableMixin, ABC):
         if self.structure is not None:
             self.structure.add_task(self)
 
-    @property
+    @cached_property
     @abstractmethod
     def input(self) -> BaseArtifact: ...
 
@@ -137,6 +138,7 @@ class BaseTask(FuturesExecutorMixin, SerializableMixin, ABC):
         return self.state == BaseTask.State.EXECUTING
 
     def before_run(self) -> None:
+        RunnableMixin.before_run(self)
         if self.structure is not None:
             EventBus.publish_event(
                 StartTaskEvent(
@@ -148,25 +150,13 @@ class BaseTask(FuturesExecutorMixin, SerializableMixin, ABC):
                 ),
             )
 
-    def after_run(self) -> None:
-        if self.structure is not None:
-            EventBus.publish_event(
-                FinishTaskEvent(
-                    task_id=self.id,
-                    task_parent_ids=self.parent_ids,
-                    task_child_ids=self.child_ids,
-                    task_input=self.input,
-                    task_output=self.output,
-                ),
-            )
-
-    def execute(self) -> Optional[BaseArtifact]:
+    def run(self) -> BaseArtifact:
         try:
             self.state = BaseTask.State.EXECUTING
 
             self.before_run()
 
-            self.output = self.run()
+            self.output = self.try_run()
 
             self.after_run()
         except Exception as e:
@@ -178,17 +168,32 @@ class BaseTask(FuturesExecutorMixin, SerializableMixin, ABC):
 
         return self.output
 
-    def can_execute(self) -> bool:
+    def after_run(self) -> None:
+        RunnableMixin.after_run(self)
+        if self.structure is not None:
+            EventBus.publish_event(
+                FinishTaskEvent(
+                    task_id=self.id,
+                    task_parent_ids=self.parent_ids,
+                    task_child_ids=self.child_ids,
+                    task_input=self.input,
+                    task_output=self.output,
+                ),
+            )
+
+    def can_run(self) -> bool:
         return self.state == BaseTask.State.PENDING and all(parent.is_finished() for parent in self.parents)
 
     def reset(self) -> BaseTask:
         self.state = BaseTask.State.PENDING
+        if self.input:
+            del self.input
         self.output = None
 
         return self
 
     @abstractmethod
-    def run(self) -> BaseArtifact: ...
+    def try_run(self) -> BaseArtifact: ...
 
     @property
     def full_context(self) -> dict[str, Any]: