+++ griptape/tasks/variation_image_generation_task.py
@@ -56,7 +56,7 @@ class VariationImageGenerationTask(BaseImageGenerationTask):
     def input(self, value: tuple[str | TextArtifact, ImageArtifact] | Callable[[BaseTask], ListArtifact]) -> None:
         self._input = value
 
-    def run(self) -> ImageArtifact:
+    def try_run(self) -> ImageArtifact:
         prompt_artifact = self.input[0]
 
         image_artifact = self.input[1]